Separate out types in har-remix. Extract HAR interfaces to own module so it can be shared with CLI. Remove next since it made the server stateful and we had no means of resetting the sequence
