--- packages/har-remix/.gitignore
@@ -1,5 +1,0 @@
-/node_modules/
-/archives/
-/dist/
-*.log
-*.tgz
