--- packages/har-remix/.vscode/tasks.json
@@ -1,19 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
-  "version": "0.1.0",
-  "command": "npm",
-  "isShellCommand": true,
-  "showOutput": "always",
-  "suppressTaskName": true,
-  "tasks": [
-    {
-      "taskName": "build",
-      "args": [
-        "run", "build"
-      ],
-      "isBuildCommand": true,
-      "problemMatcher": "$tsc"
-    }
-  ]
-}
