+++ packages/har-remix/package.json
@@ -20,19 +20,18 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc",
-    "generate-har": "node scripts/generate_har_interfaces.js",
     "lint": "tslint -p tsconfig.json",
     "prepare": "tsc",
     "test": "echo \"TODO add tests for har-remix\""
   },
   "dependencies": {
+    "@tracerbench/har": "^2.0.0-beta.3",
     "mime-types": "^2.1.20"
   },
   "devDependencies": {
-    "@krisselden/har-schema": "^0.1.0",
     "@types/mime-types": "^2.1.0",
     "@types/node": "^10.12.18",
-    "json-schema-to-typescript": "^5.5.0",
+    "json-schema-to-typescript": "^6.1.3",
     "prettier": "^1.16.4",
     "tslint": "^5.14.0",
     "tslint-config-prettier": "^1.18.0",
